requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django~=3.2.18
django-allauth~=0.43.0
django-environ==0.4.5
django-model-utils==4.0.0
django-psycopg2-extension==0.0.4
gunicorn==20.0.4
idna==2.10
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysql-connector-python==8.0.21
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pylint==2.5.3
PyMySQL==0.10.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
djangorestframework~=3.14.0